combined tests for jacobian and hessian
